Removed unwanted files and folders
