--- src/front/js/component/navbar/navbarComponents/LogoDisplay.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const LogoDisplay = () => {
-    return (
-        <div>este es el LogoDisplay</div>
-    )
-}
