--- src/front/js/component/navbar/navbarComponents/ProfileDropdown.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const ProfileDropdown = () => {
-    return (
-        <div>ProfileDropdown</div>
-    )
-}
