--- src/front/js/component/navbar/navbarComponents/SingComponent.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const SingComponent = () => {
-    return (
-        <div>SingComponent</div>
-    )
-}
